try to fix it

+++ include/aws/http/private/h2_connection.h
@@ -99,8 +99,8 @@ struct aws_h2_connection {
          * connection */
         size_t window_size_self;
         /* The size dropped */
-        size_t window_size_self_dropped;
-        size_t window_size_self_dropped_threshold;
+        uint32_t window_size_self_dropped;
+        uint32_t window_size_self_dropped_threshold;
 
         /* Highest self-initiated stream-id that peer might have processed.
          * Defaults to max stream-id, may be lowered when GOAWAY frame received. */
